+++ src/BiliLite.UWP/BiliLite.UWP.csproj
@@ -153,7 +153,35 @@
     <Compile Include="Models\Common\Live\LiveRoomSuperChatModel.cs" />
     <Compile Include="Models\Common\Live\LiveRoomSuperChatUserInfoModel.cs" />
     <Compile Include="Models\Common\Live\LiveAnchorProfile.cs" />
-    <Compile Include="Modules\Detail\LiveAnchorProfilePkInfo.cs" />
+    <Compile Include="Models\Common\Live\LiveAnchorInfoBaseInfoModel.cs" />
+    <Compile Include="Models\Common\Live\LiveAnchorInfoLiveInfoModel.cs" />
+    <Compile Include="Models\Common\Live\LiveAnchorInfoModel.cs" />
+    <Compile Include="Models\Common\Live\LiveAnchorInfoOfficialInfoModel.cs" />
+    <Compile Include="Models\Common\Live\LiveAnchorInfoRelationInfoModel.cs" />
+    <Compile Include="Models\Common\Live\LiveAnchorProfileGloryInfo.cs" />
+    <Compile Include="Models\Common\Live\LiveAnchorProfilePkInfo.cs" />
+    <Compile Include="Models\Common\Live\LiveBagGiftItem.cs" />
+    <Compile Include="Models\Common\Live\LiveGiftItem.cs" />
+    <Compile Include="Models\Common\Live\LiveGiftItemCountMap.cs" />
+    <Compile Include="Models\Common\Live\LiveInfoModel.cs" />
+    <Compile Include="ViewModels\Live\LiveRoomAnchorLotteryViewModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomGiftItem.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomGuardInfoModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomInfoModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomInfoPendantsFrameModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomInfoPendantsModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomPlayUrlModel.cs" />
+    <Compile Include="ViewModels\Live\LiveRoomRankViewModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomRealPlayUrlsModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomWebUrlCodecItemModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomWebUrlFormatItemModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomWebUrlPlayurlInfoItemModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomWebUrlPlayurlItemModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomWebUrlQualityDescriptionItemModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomWebUrlStreamItemModel.cs" />
+    <Compile Include="Models\Common\Live\LiveRoomWebUrlUrlinfoItemModel.cs" />
+    <Compile Include="Models\Common\Live\LiveTitleModel.cs" />
+    <Compile Include="Models\Common\Live\LiveWalletInfo.cs" />
     <Compile Include="ViewModels\UserDynamic\UserDynamicItemDisplayViewModel.cs" />
     <Compile Include="Models\Common\UserDynamic\UserDynamicItemDisplayOneRowInfo.cs" />
     <Compile Include="Models\Common\UserDynamic\UserDynamicItemDisplayImageInfo.cs" />
@@ -495,7 +523,7 @@
     <Compile Include="Modules\Home\HomeVM.cs" />
     <Compile Include="Modules\Home\HotVM.cs" />
     <Compile Include="Modules\IModules.cs" />
-    <Compile Include="Modules\Detail\LiveRoomVM.cs" />
+    <Compile Include="ViewModels\Live\LiveRoomViewModel.cs" />
     <Compile Include="Modules\Home\LiveVM.cs" />
     <Compile Include="ViewModels\UserDynamic\UserDynamicRepostViewModel.cs" />
     <Compile Include="ViewModels\UserDynamic\UserDynamicViewModel.cs" />
