+++ src/BiliLite.UWP/Models/Common/Live/LiveAnchorInfoLiveInfoModel.cs
@@ -1,0 +1,12 @@
+﻿using Newtonsoft.Json;
+
+namespace BiliLite.Models.Common.Live
+{
+    public class LiveAnchorInfoLiveInfoModel
+    {
+        public int Level { get; set; }
+
+        [JsonProperty("level_color")]
+        public int LevelColor { get; set; }
+    }
+}
