+++ src/BiliLite.UWP/Models/Common/Live/LiveAnchorInfoOfficialInfoModel.cs
@@ -1,0 +1,11 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveAnchorInfoOfficialInfoModel
+    {
+        public int Role { get; set; }
+
+        public string Title { get; set; }
+
+        public string Desc { get; set; }
+    }
+}
