+++ src/BiliLite.UWP/Models/Common/Live/LiveAnchorInfoRelationInfoModel.cs
@@ -1,0 +1,7 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveAnchorInfoRelationInfoModel
+    {
+        public int Attention { get; set; }
+    }
+}
