+++ src/BiliLite.UWP/Models/Common/Live/LiveGiftItemCountMap.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace BiliLite.Models.Common.Live
+{
+    [Serializable]
+    public class LiveGiftItemCountMap
+    {
+        public int Num { get; set; }
+
+        public string Text { get; set; }
+    }
+}
