+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomGuardInfoModel.cs
@@ -1,0 +1,12 @@
+﻿using Newtonsoft.Json;
+
+namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomGuardInfoModel
+    {
+        public int Count { get; set; }
+
+        [JsonProperty("achievement_level")]
+        public int AchievementLevel { get; set; }
+    }
+}
