+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomInfoPendantsFrameModel.cs
@@ -1,0 +1,13 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomInfoPendantsFrameModel
+    {
+        public string Name { get; set; }
+
+        public int Position { get; set; }
+
+        public string Value { get; set; }
+
+        public string Desc { get; set; }
+    }
+}
