+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomInfoPendantsModel.cs
@@ -1,0 +1,9 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomInfoPendantsModel
+    {
+        public LiveRoomInfoPendantsFrameModel Frame { get; set; }
+
+        public object Badge { get; set; }
+    }
+}
