+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomPlayUrlModel.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomPlayUrlModel
+    {
+        [JsonProperty("playurl_info")]
+        public LiveRoomWebUrlPlayurlInfoItemModel PlayUrlInfo { get; set; }
+    }
+}
