+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomRealPlayUrlsModel.cs
@@ -1,0 +1,9 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomRealPlayUrlsModel
+    {
+        public string Name { get; set; }
+
+        public string Url { get; set; }
+    }
+}
