+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomSuperChatUserInfoModel.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace BiliLite.Modules.LiveRoomDetailModels
+namespace BiliLite.Models.Common.Live
 {
     public class LiveRoomSuperChatUserInfoModel
     {
