+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomWebUrlFormatItemModel.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomWebUrlFormatItemModel
+    {
+        public List<LiveRoomWebUrlCodecItemModel> Codec { get; set; }
+    }
+}
