+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomWebUrlPlayurlInfoItemModel.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomWebUrlPlayurlInfoItemModel
+    {
+        [JsonProperty("playurl")]
+        public LiveRoomWebUrlPlayurlItemModel PlayUrl { get; set; }
+    }
+}
