+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomWebUrlPlayurlItemModel.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomWebUrlPlayurlItemModel
+    {
+        [JsonProperty("g_qn_desc")]
+        public List<LiveRoomWebUrlQualityDescriptionItemModel> GQnDesc { get; set; }
+
+        public List<LiveRoomWebUrlStreamItemModel> Stream { get; set; }
+    }
+}
