+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomWebUrlQualityDescriptionItemModel.cs
@@ -1,0 +1,9 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomWebUrlQualityDescriptionItemModel
+    {
+        public int Qn { get; set; }
+
+        public string Desc { get; set; }
+    }
+}
