+++ src/BiliLite.UWP/Models/Common/Live/LiveRoomWebUrlUrlinfoItemModel.cs
@@ -1,0 +1,11 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveRoomWebUrlUrlinfoItemModel
+    {
+        public string Name { get; set; }
+
+        public string Host { get; set; }
+
+        public string Extra { get; set; }
+    }
+}
