+++ src/BiliLite.UWP/Models/Common/Live/LiveTitleModel.cs
@@ -1,0 +1,11 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveTitleModel
+    {
+        public string Id { get; set; }
+
+        public string Title { get; set; }
+
+        public string Img { get; set; }
+    }
+}
