+++ src/BiliLite.UWP/Models/Common/Live/LiveWalletInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace BiliLite.Models.Common.Live
+{
+    public class LiveWalletInfo
+    {
+        public int Gold { get; set; }
+
+        public int Silver { get; set; }
+    }
+}
